rename acorn sources with .mjs extension
