+++ acorn/src/regexp.mjs
@@ -1,7 +1,7 @@
-import {isIdentifierStart, isIdentifierChar} from "./identifier.js"
-import {Parser} from "./state.js"
-import UNICODE_PROPERTY_VALUES from "./unicode-property-data.js"
-import {has} from "./util.js"
+import {isIdentifierStart, isIdentifierChar} from "./identifier.mjs"
+import {Parser} from "./state.mjs"
+import UNICODE_PROPERTY_VALUES from "./unicode-property-data.mjs"
+import {has} from "./util.mjs"
 
 const pp = Parser.prototype
