+++ acorn/src/state.mjs
@@ -1,9 +1,9 @@
-import {reservedWords, keywords} from "./identifier"
-import {types as tt} from "./tokentype"
-import {lineBreak} from "./whitespace"
-import {getOptions} from "./options"
-import {wordsRegexp} from "./util"
-import {SCOPE_TOP, SCOPE_FUNCTION, SCOPE_ASYNC, SCOPE_GENERATOR, SCOPE_SUPER, SCOPE_DIRECT_SUPER} from "./scopeflags"
+import {reservedWords, keywords} from "./identifier.mjs"
+import {types as tt} from "./tokentype.mjs"
+import {lineBreak} from "./whitespace.mjs"
+import {getOptions} from "./options.mjs"
+import {wordsRegexp} from "./util.mjs"
+import {SCOPE_TOP, SCOPE_FUNCTION, SCOPE_ASYNC, SCOPE_GENERATOR, SCOPE_SUPER, SCOPE_DIRECT_SUPER} from "./scopeflags.mjs"
 
 export class Parser {
   constructor(options, input, startPos) {
