+++ acorn/src/util.mjs
